Add UserController for saving user data and record in Firestore
This commit adds a new file `user_controller.dart` which contains the `UserController` class. This class has two functions: `saveUserData` and `saveUserRecordInFirestore` which are responsible for saving user data and record in Firestore respectively. These functions use the `SaveUserDataUseCase` and `SaveUserRecordInFirestoreUseCase` classes to perform the actual saving. The commit also imports necessary dependencies and handles exceptions with error messages.