Refactor dependency injection in app module
This commit refactors the dependency injection in the app module by adding new use cases for updating user details, uploading images, and re-authenticating email and password. It also registers the new use case for fetching user details and updates the registration of the existing use case for saving user data. Additionally, the formatting of the code has been improved for better readability.