Add AppShimmerEffectWidget
This commit adds the AppShimmerEffectWidget to the project. This widget is used to display a shimmer effect on a container with customizable width, height, and radius. The color of the container can also be customized or it will default to either AppColors.darkerGrey or AppColors.white depending on the current theme.