Refactor Personalization profile pages
This commit refactors the Personalization profile pages by reorganizing imports and removing unused code. Specifically, the `sizes.dart` import is moved to the top and the `custom_app_bar.dart` import is moved to the middle of the `change_name_screen.dart` file. In addition, the `re_auth_form_widget.dart` import is moved to the bottom of the `re_auth_login_form_screen.dart` file.