Add factory method to BrandModel for JSON deserialization
This commit adds a new factory method to the BrandModel class that allows for deserialization of JSON data. The new method, `fromJ`, takes a Map of String to dynamic and returns a new instance of BrandModel with the data from the map. If the map is empty, it returns an empty instance of BrandModel. This will be useful for parsing JSON data when retrieving brand information from the server.