Refactor product card widgets and add shimmer effect
This commit refactors the product card widgets to use the `ProductEntity` model and adds a shimmer effect to the `VerticalProductShimmerWidget`. It also adds a `price` parameter to the `PriceAndAddToCartWidget` and updates the `ProductCardHorizontalWidget` and `ProductCardVerticalWidget` to use it. Additionally, the `ProductCardVerticalWidget` now navigates to the `ProductDetailsScreen` with the `ProductEntity` as an argument. Finally, the `SortableProducts` widget now uses an empty `ProductEntity` model for the `ProductCardVerticalWidget`.